+++ src/temp/PermissionSetDetails.js
@@ -14,6 +14,7 @@ import {
 import { ViewMetaData } from '@folio/stripes/smart-components';
 
 import RenderPermissions from './RenderPermissions';
+import AssignedUsers from './AssignedUsers';
 
 class PermissionSetDetails extends React.Component {
   static propTypes = {
@@ -21,6 +22,7 @@ class PermissionSetDetails extends React.Component {
       connect: PropTypes.func.isRequired,
     }).isRequired,
     initialValues: PropTypes.object,
+    tenantId: PropTypes.string.isRequired,
   };
 
   constructor(props) {
@@ -31,7 +33,8 @@ class PermissionSetDetails extends React.Component {
     this.state = {
       sections: {
         generalInformation: true,
-        assignedPermissions: true,
+        assignedPermissions: false,
+        assignedUsers: true,
       },
     };
 
@@ -59,7 +62,7 @@ class PermissionSetDetails extends React.Component {
   }
 
   render() {
-    const { initialValues: selectedSet } = this.props;
+    const { initialValues: selectedSet, tenantId } = this.props;
     const { sections } = this.state;
     const untitledPermissionSetMessage = <FormattedMessage id="ui-users.permissions.untitledPermissionSet" />;
 
@@ -122,6 +125,12 @@ class PermissionSetDetails extends React.Component {
           permToDelete="perms.permissions.item.delete"
           permToModify="perms.permissions.item.put"
           {...this.props}
+        />
+        <AssignedUsers
+          onToggle={this.handleSectionToggle}
+          expanded={sections.assignedUsers}
+          permissionSetId={selectedSet.id}
+          tenantId={tenantId}
         />
       </div>
     );